flask-dummy/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
